The am335x_evm.sh script will, through autoconfigure_usb0.sh, read the network-values for usb0 from /etc/network/interfaces, instead of having them hardcoded into the script itself.
If udhcpd is installed, changing /etc/network/interfaces will cause
the scripts to reconfigure /etc/udhcpd.conf (provided that the udhcpd
interface is set to usb0).

If dnsmasq is installed (and udhcpd is not), changing
/etc/network/interfaces will cause the scripts to reconfigure
/etc/dnsmasq.d/usb0-dhcp.generated accordingly, only if no other
configuration file references usb0.  The usb0-dhcp.generated filename
can be appended with ~ (tilde) to disable further reconfiguration.

Signed-off-by: Per Dalgas Jakobsen <pdj@knaldgas.dk>