Comments out unmounting /tmp.
Filesystems mount on subdirectories of /tmp are already unmounted earlier
in the script. /tmp itself is not a mount point in all systems, and if it
isn't, umount will fail when called to unmount it, resulting in premature
termination of the flasher script. Also, it doesn't seem necessary to
unmount /tmp in this script (which doesn't run as init) even if it is a
mounted filesystem.