Add CDC/ECM function to USB gadget
This adds a CDC/EMC USB networking function to the multifunction
USB gadget device. This is needed to get USB networking working on
macOS with no 3rd party drivers. It also works better than RNDIS on
Linux. Also, we will be changing the USB class of the RNDIS driver,
so it will no longer be detected by Linux, so CDC will be the only
networking interface that shows up on Linux.