Add autoconfigure_usb1.sh
This adds a new boot script for configuring the usb1 network interface.
This interface is associated with the CDC/ECM function of the USB gadget
device. We are using the tethering feature of ConnMan to dish up DHCP
addresses on this interface. (udchpd only supports a single network
interface, so we can't set this up like usb0.)