Add a second USB gadget configuration for Linux
This adds a second USB gadget configuration that will be used by Linux.
Linux is smart enough to know that RNDIS is not nice, so if it sees RNDIS
as the first interface and it sees that there is a second USB configuration,
it will prefer the second configuration. This results in only the CDC ECM
interface being used on Linux (and older versions of macOS).